checkers/requirements.txt
websocket_client
requests
checklib
pycrypto